refactor(ui): extract workflow TEMPLATES data out of TemplateGallery
TemplateGallery.tsx was 1379 lines, ~1240 of which were the pure TEMPLATES
data array. Move the WorkflowTemplate type + TEMPLATES catalog into
workflow-templates.ts; TemplateGallery.tsx drops to 122 lines (the actual
component). The component re-exports WorkflowTemplate so the workflows
barrel is unaffected. Pure data move — behaviour unchanged.

UI typecheck + lint + 347 tests + build green.
