+++ packages/ui/src/components/workflows/workflow-templates.ts
@@ -1,0 +1,1268 @@
+/**
+ * Workflow template catalog.
+ *
+ * Pure data extracted from TemplateGallery.tsx — the built-in workflow
+ * templates (copilot definition format, converted via
+ * convertDefinitionToReactFlow) plus their shared type.
+ */
+
+export interface WorkflowTemplate {
+  id: string;
+  name: string;
+  description: string;
+  category: string;
+  nodeCount: number;
+  definition: {
+    name: string;
+    nodes: unknown[];
+    edges: unknown[];
+    variables?: Record<string, unknown>;
+  };
+}
+
+// Templates use copilot definition format — converted via convertDefinitionToReactFlow.
+// LLM nodes use provider/model 'default' — resolved at runtime to the user's configured defaults.
+// Tool nodes use full dotted names (core.search_memories, core.list_goals, etc.).
+export const TEMPLATES: WorkflowTemplate[] = [
+  {
+    id: 'daily-memory-briefing',
+    name: 'Daily Memory Briefing',
+    description: 'Search recent memories, summarize with AI, and deliver a morning briefing',
+    category: 'Scheduling',
+    nodeCount: 4,
+    definition: {
+      name: 'Daily Memory Briefing',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'schedule',
+          label: 'Daily 9AM',
+          cron: '0 9 * * *',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          tool: 'core.search_memories',
+          label: 'Search Recent Memories',
+          args: { query: 'important events tasks updates', limit: 20 },
+          position: { x: 350, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'llm',
+          label: 'Summarize Briefing',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            "You are a concise personal assistant. Create a clear morning briefing from the user's recent memories. Group by topic, highlight action items, and keep it scannable.",
+          userMessage: 'Create my daily briefing from these recent memories:\n\n{{node_2.output}}',
+          temperature: 0.5,
+          maxTokens: 2048,
+          position: { x: 600, y: 150 },
+        },
+        {
+          id: 'node_4',
+          type: 'notification',
+          label: 'Deliver Briefing',
+          message: '{{node_3.output}}',
+          severity: 'info',
+          position: { x: 850, y: 150 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4' },
+      ],
+    },
+  },
+  {
+    id: 'goal-progress-tracker',
+    name: 'Goal Progress Tracker',
+    description: 'Review active goals, analyze progress with AI, flag blocked items',
+    category: 'Automation',
+    nodeCount: 6,
+    definition: {
+      name: 'Goal Progress Tracker',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'schedule',
+          label: 'Weekly Monday 8AM',
+          cron: '0 8 * * 1',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          tool: 'core.list_goals',
+          label: 'Get Active Goals',
+          args: { status: 'active', limit: 50 },
+          position: { x: 300, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'llm',
+          label: 'Analyze Progress',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            "You are a productivity coach. Analyze the user's goals and their steps. Identify: (1) goals making good progress, (2) goals that are stalled or blocked, (3) suggested next actions. Be specific and actionable.",
+          userMessage: 'Analyze my goal progress:\n\n{{node_2.output}}',
+          temperature: 0.4,
+          maxTokens: 3000,
+          position: { x: 550, y: 150 },
+        },
+        {
+          id: 'node_4',
+          type: 'condition',
+          label: 'Any Blocked?',
+          expression:
+            'typeof data === "string" && (data.toLowerCase().includes("blocked") || data.toLowerCase().includes("stalled"))',
+          position: { x: 800, y: 150 },
+        },
+        {
+          id: 'node_5',
+          type: 'notification',
+          label: 'Alert: Blocked Goals',
+          message: '{{node_3.output}}',
+          severity: 'warning',
+          position: { x: 1050, y: 50 },
+        },
+        {
+          id: 'node_6',
+          type: 'notification',
+          label: 'Weekly Update',
+          message: '{{node_3.output}}',
+          severity: 'info',
+          position: { x: 1050, y: 250 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4' },
+        { source: 'node_4', target: 'node_5', sourceHandle: 'true' },
+        { source: 'node_4', target: 'node_6', sourceHandle: 'false' },
+      ],
+    },
+  },
+  {
+    id: 'smart-content-pipeline',
+    name: 'Smart Content Pipeline',
+    description: 'Generate content with AI, auto-review for quality, revise if needed',
+    category: 'AI',
+    nodeCount: 6,
+    definition: {
+      name: 'Smart Content Pipeline',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'manual',
+          label: 'Start',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          type: 'llm',
+          label: 'Generate Draft',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            "You are an expert content writer. Write high-quality, engaging content based on the user's request.",
+          userMessage:
+            'Write a {{variables.contentType}} about: {{variables.topic}}\n\nTone: {{variables.tone}}\nTarget audience: {{variables.audience}}',
+          temperature: 0.8,
+          maxTokens: 4096,
+          position: { x: 350, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'llm',
+          label: 'Quality Review',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a strict content reviewer. Rate the content on a scale of 1-10 for quality, clarity, and engagement. If the score is 7 or above, respond with "PASS: [score]". If below 7, respond with "REVISE: [specific improvements needed]".',
+          userMessage: 'Review this {{variables.contentType}}:\n\n{{node_2.output}}',
+          temperature: 0.3,
+          maxTokens: 1024,
+          position: { x: 600, y: 150 },
+        },
+        {
+          id: 'node_4',
+          type: 'condition',
+          label: 'Quality Check',
+          expression: 'typeof data === "string" && data.startsWith("PASS")',
+          position: { x: 850, y: 150 },
+        },
+        {
+          id: 'node_5',
+          type: 'notification',
+          label: 'Content Ready',
+          message: 'Your {{variables.contentType}} is ready:\n\n{{node_2.output}}',
+          severity: 'success',
+          position: { x: 1100, y: 50 },
+        },
+        {
+          id: 'node_6',
+          type: 'llm',
+          label: 'Revise Content',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            "You are an expert editor. Revise the content based on the reviewer's feedback while preserving the original intent and voice.",
+          userMessage:
+            'Original content:\n{{node_2.output}}\n\nReviewer feedback:\n{{node_3.output}}\n\nPlease revise accordingly.',
+          temperature: 0.6,
+          maxTokens: 4096,
+          position: { x: 1100, y: 250 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4' },
+        { source: 'node_4', target: 'node_5', sourceHandle: 'true' },
+        { source: 'node_4', target: 'node_6', sourceHandle: 'false' },
+      ],
+      variables: {
+        contentType: 'blog post',
+        topic: 'AI automation in daily life',
+        tone: 'professional but approachable',
+        audience: 'tech-savvy professionals',
+      },
+    },
+  },
+  {
+    id: 'webhook-processor',
+    name: 'Webhook Processor',
+    description: 'Receive webhook, validate payload, classify with AI, and route to handlers',
+    category: 'Integration',
+    nodeCount: 7,
+    definition: {
+      name: 'Webhook Processor',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'webhook',
+          label: 'Incoming Webhook',
+          webhookPath: '/events',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          type: 'code',
+          label: 'Validate Payload',
+          language: 'javascript',
+          code: 'const body = typeof data === "string" ? JSON.parse(data) : (data.body || data);\nif (!body || !body.type) throw new Error("Invalid payload: missing type field");\nreturn { type: body.type, payload: body, receivedAt: new Date().toISOString() };',
+          position: { x: 350, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'switch',
+          label: 'Route by Type',
+          expression: 'data.type',
+          cases: [
+            { label: 'user_signup', value: 'user_signup' },
+            { label: 'payment', value: 'payment' },
+            { label: 'error', value: 'error' },
+          ],
+          position: { x: 600, y: 150 },
+        },
+        {
+          id: 'node_4',
+          tool: 'core.create_memory',
+          label: 'Save New User',
+          args: { content: 'New user signed up: {{node_2.output.payload}}', importance: 7 },
+          position: { x: 900, y: 30 },
+        },
+        {
+          id: 'node_5',
+          tool: 'core.add_custom_record',
+          label: 'Log Payment',
+          args: { table: 'payments', data: '{{node_2.output.payload}}' },
+          position: { x: 900, y: 150 },
+        },
+        {
+          id: 'node_6',
+          type: 'notification',
+          label: 'Error Alert',
+          message: 'Webhook error event received:\n{{node_2.output.payload}}',
+          severity: 'error',
+          position: { x: 900, y: 270 },
+        },
+        {
+          id: 'node_7',
+          type: 'notification',
+          label: 'Unknown Type',
+          message: 'Unhandled webhook type: {{node_2.output.type}}',
+          severity: 'warning',
+          position: { x: 900, y: 390 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4', sourceHandle: 'user_signup' },
+        { source: 'node_3', target: 'node_5', sourceHandle: 'payment' },
+        { source: 'node_3', target: 'node_6', sourceHandle: 'error' },
+        { source: 'node_3', target: 'node_7', sourceHandle: 'default' },
+      ],
+    },
+  },
+  {
+    id: 'task-digest',
+    name: 'Weekly Task Digest',
+    description: 'Gather tasks, goals, and memories into a comprehensive weekly digest',
+    category: 'Scheduling',
+    nodeCount: 7,
+    definition: {
+      name: 'Weekly Task Digest',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'schedule',
+          label: 'Friday 5PM',
+          cron: '0 17 * * 5',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          type: 'parallel',
+          label: 'Gather Data',
+          branchCount: 3,
+          branchLabels: ['Tasks', 'Goals', 'Memories'],
+          position: { x: 300, y: 150 },
+        },
+        {
+          id: 'node_3',
+          tool: 'core.list_tasks',
+          label: 'Get Tasks',
+          args: { limit: 30 },
+          position: { x: 550, y: 30 },
+        },
+        {
+          id: 'node_4',
+          tool: 'core.list_goals',
+          label: 'Get Goals',
+          args: { status: 'active', limit: 20 },
+          position: { x: 550, y: 150 },
+        },
+        {
+          id: 'node_5',
+          tool: 'core.search_memories',
+          label: 'Get Key Memories',
+          args: { query: 'achievements progress milestones completed', limit: 15 },
+          position: { x: 550, y: 270 },
+        },
+        {
+          id: 'node_6',
+          type: 'merge',
+          label: 'Combine All',
+          mode: 'waitAll',
+          position: { x: 800, y: 150 },
+        },
+        {
+          id: 'node_7',
+          type: 'llm',
+          label: 'Format Digest',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a personal productivity assistant. Create a well-structured weekly digest with sections for: completed tasks, pending tasks, goal progress, and key highlights. Use bullet points and be concise.',
+          userMessage:
+            'Format my weekly digest from this data:\n\nTasks: {{node_3.output}}\n\nGoals: {{node_4.output}}\n\nKey Memories: {{node_5.output}}',
+          temperature: 0.5,
+          maxTokens: 3000,
+          position: { x: 1050, y: 150 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3', sourceHandle: 'branch-0' },
+        { source: 'node_2', target: 'node_4', sourceHandle: 'branch-1' },
+        { source: 'node_2', target: 'node_5', sourceHandle: 'branch-2' },
+        { source: 'node_3', target: 'node_6' },
+        { source: 'node_4', target: 'node_6' },
+        { source: 'node_5', target: 'node_6' },
+        { source: 'node_6', target: 'node_7' },
+      ],
+    },
+  },
+  {
+    id: 'approval-flow',
+    name: 'Approval Flow',
+    description: 'AI risk analysis, human approval gate, then execute with audit trail',
+    category: 'Automation',
+    nodeCount: 7,
+    definition: {
+      name: 'Approval Flow',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'manual',
+          label: 'Submit Request',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          type: 'llm',
+          label: 'Risk Analysis',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a risk analyst. Evaluate the request and classify risk as LOW, MEDIUM, or HIGH. Provide a brief justification. Format: "RISK: [level]\\n\\nAnalysis: [details]"',
+          userMessage: 'Evaluate this request:\n\n{{node_1.output}}',
+          temperature: 0.3,
+          maxTokens: 1024,
+          position: { x: 350, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'approval',
+          label: 'Manager Approval',
+          approvalMessage:
+            'Request: {{node_1.output}}\n\nAI Risk Assessment:\n{{node_2.output}}\n\nPlease approve or reject.',
+          timeoutMinutes: 1440,
+          position: { x: 600, y: 150 },
+        },
+        {
+          id: 'node_4',
+          tool: 'core.create_memory',
+          label: 'Log Approval',
+          args: {
+            content: 'Request approved: {{node_1.output}}. Risk assessment: {{node_2.output}}',
+            importance: 8,
+          },
+          position: { x: 850, y: 50 },
+        },
+        {
+          id: 'node_5',
+          type: 'notification',
+          label: 'Approved',
+          message: 'Your request has been approved and logged.',
+          severity: 'success',
+          position: { x: 1100, y: 50 },
+        },
+        {
+          id: 'node_6',
+          tool: 'core.create_memory',
+          label: 'Log Rejection',
+          args: { content: 'Request rejected: {{node_1.output}}', importance: 5 },
+          position: { x: 850, y: 250 },
+        },
+        {
+          id: 'node_7',
+          type: 'notification',
+          label: 'Rejected',
+          message: 'Your request was not approved.',
+          severity: 'warning',
+          position: { x: 1100, y: 250 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4', sourceHandle: 'approved' },
+        { source: 'node_3', target: 'node_6', sourceHandle: 'rejected' },
+        { source: 'node_4', target: 'node_5' },
+        { source: 'node_6', target: 'node_7' },
+      ],
+    },
+  },
+  {
+    id: 'multi-intent-router',
+    name: 'Multi-Intent Router',
+    description: 'Classify user input into categories and route to specialized AI handlers',
+    category: 'AI',
+    nodeCount: 7,
+    definition: {
+      name: 'Multi-Intent Router',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'manual',
+          label: 'User Input',
+          position: { x: 100, y: 200 },
+        },
+        {
+          id: 'node_2',
+          type: 'llm',
+          label: 'Classify Intent',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'Classify the user input into exactly one category. Reply with ONLY the category name, nothing else.\n\nCategories:\n- question (factual questions, how-to)\n- task (action items, things to do)\n- creative (writing, brainstorming, ideas)\n- analysis (data interpretation, research)',
+          userMessage: '{{node_1.output}}',
+          temperature: 0.1,
+          maxTokens: 50,
+          position: { x: 350, y: 200 },
+        },
+        {
+          id: 'node_3',
+          type: 'switch',
+          label: 'Route Intent',
+          expression: 'typeof data === "string" ? data.trim().toLowerCase() : "question"',
+          cases: [
+            { label: 'question', value: 'question' },
+            { label: 'task', value: 'task' },
+            { label: 'creative', value: 'creative' },
+            { label: 'analysis', value: 'analysis' },
+          ],
+          position: { x: 600, y: 200 },
+        },
+        {
+          id: 'node_4',
+          type: 'llm',
+          label: 'Answer Question',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a knowledgeable assistant. Provide clear, accurate answers with sources when possible.',
+          userMessage: '{{node_1.output}}',
+          temperature: 0.4,
+          maxTokens: 2048,
+          position: { x: 900, y: 30 },
+        },
+        {
+          id: 'node_5',
+          type: 'llm',
+          label: 'Plan Task',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a task planner. Break down the request into clear, actionable steps with priorities and estimated effort.',
+          userMessage: 'Create an action plan for: {{node_1.output}}',
+          temperature: 0.5,
+          maxTokens: 2048,
+          position: { x: 900, y: 150 },
+        },
+        {
+          id: 'node_6',
+          type: 'llm',
+          label: 'Creative Work',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a creative writing assistant. Be imaginative, original, and engaging.',
+          userMessage: '{{node_1.output}}',
+          temperature: 0.9,
+          maxTokens: 4096,
+          position: { x: 900, y: 270 },
+        },
+        {
+          id: 'node_7',
+          type: 'llm',
+          label: 'Deep Analysis',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a research analyst. Provide thorough, structured analysis with pros/cons, key findings, and recommendations.',
+          userMessage: 'Analyze: {{node_1.output}}',
+          temperature: 0.4,
+          maxTokens: 4096,
+          position: { x: 900, y: 390 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4', sourceHandle: 'question' },
+        { source: 'node_3', target: 'node_5', sourceHandle: 'task' },
+        { source: 'node_3', target: 'node_6', sourceHandle: 'creative' },
+        { source: 'node_3', target: 'node_7', sourceHandle: 'analysis' },
+      ],
+    },
+  },
+  {
+    id: 'data-import-pipeline',
+    name: 'Data Import Pipeline',
+    description: 'Fetch API data, validate each record, and store in custom data tables',
+    category: 'Data',
+    nodeCount: 7,
+    definition: {
+      name: 'Data Import Pipeline',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'manual',
+          label: 'Start Import',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          type: 'httpRequest',
+          label: 'Fetch API Data',
+          method: 'GET',
+          url: '{{variables.apiUrl}}',
+          headers: { Authorization: 'Bearer {{variables.apiToken}}' },
+          position: { x: 300, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'code',
+          label: 'Parse & Validate',
+          language: 'javascript',
+          code: 'const response = data.body || data;\nconst items = Array.isArray(response) ? response : (response.data || response.results || []);\nconst valid = items.filter(item => {\n  if (!item || typeof item !== "object") return false;\n  if (!item.id && !item.name) return false;\n  return true;\n});\nreturn { total: items.length, valid: valid.length, skipped: items.length - valid.length, records: valid };',
+          position: { x: 550, y: 150 },
+        },
+        {
+          id: 'node_4',
+          type: 'condition',
+          label: 'Has Records?',
+          expression: 'data && data.valid > 0',
+          position: { x: 800, y: 150 },
+        },
+        {
+          id: 'node_5',
+          type: 'forEach',
+          label: 'Store Each Record',
+          arrayExpression: '{{node_3.output.records}}',
+          itemVariable: 'record',
+          maxIterations: 500,
+          onError: 'continue',
+          position: { x: 1050, y: 50 },
+        },
+        {
+          id: 'node_6',
+          tool: 'core.add_custom_record',
+          label: 'Save to Table',
+          args: { table: '{{variables.tableName}}', data: '{{record}}' },
+          position: { x: 1050, y: 200 },
+        },
+        {
+          id: 'node_7',
+          type: 'notification',
+          label: 'No Data',
+          message: 'API returned no valid records from {{variables.apiUrl}}',
+          severity: 'warning',
+          position: { x: 1050, y: 300 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4' },
+        { source: 'node_4', target: 'node_5', sourceHandle: 'true' },
+        { source: 'node_4', target: 'node_7', sourceHandle: 'false' },
+        { source: 'node_5', target: 'node_6', sourceHandle: 'each' },
+      ],
+      variables: {
+        apiUrl: 'https://api.example.com/data',
+        apiToken: '',
+        tableName: 'imported_records',
+      },
+    },
+  },
+  {
+    id: 'knowledge-capture',
+    name: 'Knowledge Capture',
+    description: 'Process input with AI, extract key facts, and save as structured memories',
+    category: 'AI',
+    nodeCount: 5,
+    definition: {
+      name: 'Knowledge Capture',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'manual',
+          label: 'Input Knowledge',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          type: 'llm',
+          label: 'Extract Key Facts',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'Extract key facts, insights, and actionable items from the input. Return a JSON array of objects, each with: {"fact": "...", "importance": 1-10, "category": "insight|action|reference"}. Return ONLY valid JSON, no explanation.',
+          userMessage: '{{node_1.output}}',
+          temperature: 0.3,
+          maxTokens: 2048,
+          position: { x: 350, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'code',
+          label: 'Parse Facts',
+          language: 'javascript',
+          code: 'const text = typeof data === "string" ? data : JSON.stringify(data);\nconst jsonMatch = text.match(/\\[\\s*\\{[\\s\\S]*\\}\\s*\\]/);\nif (!jsonMatch) return [];\ntry { return JSON.parse(jsonMatch[0]); } catch { return []; }',
+          position: { x: 600, y: 150 },
+        },
+        {
+          id: 'node_4',
+          type: 'forEach',
+          label: 'Save Each Fact',
+          arrayExpression: '{{node_3.output}}',
+          itemVariable: 'fact',
+          onError: 'continue',
+          position: { x: 850, y: 150 },
+        },
+        {
+          id: 'node_5',
+          tool: 'core.create_memory',
+          label: 'Store Memory',
+          args: { content: '{{fact.fact}}', importance: '{{fact.importance}}' },
+          position: { x: 850, y: 300 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4' },
+        { source: 'node_4', target: 'node_5', sourceHandle: 'each' },
+      ],
+    },
+  },
+  {
+    id: 'parallel-research',
+    name: 'Parallel Research',
+    description: 'Run 3 specialized AI analyses in parallel, merge results, synthesize a report',
+    category: 'AI',
+    nodeCount: 8,
+    definition: {
+      name: 'Parallel Research',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'manual',
+          label: 'Research Topic',
+          position: { x: 100, y: 200 },
+        },
+        {
+          id: 'node_2',
+          type: 'parallel',
+          label: 'Research Branches',
+          branchCount: 3,
+          branchLabels: ['Technical', 'Market', 'Risks'],
+          position: { x: 300, y: 200 },
+        },
+        {
+          id: 'node_3',
+          type: 'llm',
+          label: 'Technical Analysis',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a technical analyst. Provide deep technical analysis including architecture, implementation challenges, and technology stack recommendations.',
+          userMessage: 'Technical analysis of: {{variables.topic}}',
+          temperature: 0.5,
+          maxTokens: 3000,
+          position: { x: 550, y: 50 },
+        },
+        {
+          id: 'node_4',
+          type: 'llm',
+          label: 'Market Analysis',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a market analyst. Analyze market size, competition, trends, target audience, and go-to-market strategy.',
+          userMessage: 'Market analysis of: {{variables.topic}}',
+          temperature: 0.5,
+          maxTokens: 3000,
+          position: { x: 550, y: 200 },
+        },
+        {
+          id: 'node_5',
+          type: 'llm',
+          label: 'Risk Assessment',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a risk analyst. Identify potential risks, regulatory concerns, scalability issues, and mitigation strategies.',
+          userMessage: 'Risk assessment for: {{variables.topic}}',
+          temperature: 0.4,
+          maxTokens: 2000,
+          position: { x: 550, y: 350 },
+        },
+        {
+          id: 'node_6',
+          type: 'merge',
+          label: 'Combine Results',
+          mode: 'waitAll',
+          position: { x: 800, y: 200 },
+        },
+        {
+          id: 'node_7',
+          type: 'llm',
+          label: 'Synthesize Report',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are an executive analyst. Synthesize the three research reports into a cohesive executive summary with key findings, recommendations, and next steps. Use clear headings and bullet points.',
+          userMessage:
+            'Synthesize these three research reports into one executive summary:\n\nTechnical Analysis:\n{{node_3.output}}\n\nMarket Analysis:\n{{node_4.output}}\n\nRisk Assessment:\n{{node_5.output}}',
+          temperature: 0.5,
+          maxTokens: 4096,
+          position: { x: 1050, y: 200 },
+        },
+        {
+          id: 'node_8',
+          type: 'notification',
+          label: 'Report Ready',
+          message: 'Research report for "{{variables.topic}}" is complete.',
+          severity: 'success',
+          position: { x: 1300, y: 200 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3', sourceHandle: 'branch-0' },
+        { source: 'node_2', target: 'node_4', sourceHandle: 'branch-1' },
+        { source: 'node_2', target: 'node_5', sourceHandle: 'branch-2' },
+        { source: 'node_3', target: 'node_6' },
+        { source: 'node_4', target: 'node_6' },
+        { source: 'node_5', target: 'node_6' },
+        { source: 'node_6', target: 'node_7' },
+        { source: 'node_7', target: 'node_8' },
+      ],
+      variables: { topic: 'AI-powered personal assistant platforms' },
+    },
+  },
+  {
+    id: 'github-issue-triage',
+    name: 'GitHub Issue Triage',
+    description:
+      'Fetch issues via HTTP, classify priority and labels with AI, route critical issues to alerts',
+    category: 'Integration',
+    nodeCount: 6,
+    definition: {
+      name: 'GitHub Issue Triage',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'schedule',
+          label: 'Every 30 Minutes',
+          cron: '*/30 * * * *',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          type: 'httpRequest',
+          label: 'Fetch Open Issues',
+          method: 'GET',
+          url: 'https://api.github.com/repos/{{variables.owner}}/{{variables.repo}}/issues?state=open&per_page=10&sort=created&direction=desc',
+          headers: {
+            Accept: 'application/vnd.github+json',
+            Authorization: 'Bearer {{variables.githubToken}}',
+          },
+          position: { x: 350, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'llm',
+          label: 'Classify Priority',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are an issue triage assistant. For each GitHub issue, classify its priority as "critical", "high", "medium", or "low" and suggest labels. Return a JSON array of objects: [{"number": <issue_number>, "title": "...", "priority": "critical|high|medium|low", "labels": ["bug","feature",...], "reason": "..."}].',
+          userMessage:
+            'Classify these GitHub issues by priority and suggest labels:\n\n{{node_2.output.body}}',
+          temperature: 0.2,
+          maxTokens: 2048,
+          responseFormat: 'json',
+          position: { x: 600, y: 150 },
+        },
+        {
+          id: 'node_4',
+          type: 'condition',
+          label: 'Any Critical?',
+          expression:
+            'Array.isArray(data) ? data.some(i => i.priority === "critical") : (typeof data === "string" && data.includes("critical"))',
+          position: { x: 850, y: 150 },
+        },
+        {
+          id: 'node_5',
+          type: 'notification',
+          label: 'Critical Alert',
+          message:
+            'CRITICAL issues found in {{variables.owner}}/{{variables.repo}}:\n\n{{node_3.output}}',
+          severity: 'error',
+          position: { x: 1100, y: 50 },
+        },
+        {
+          id: 'node_6',
+          type: 'notification',
+          label: 'Triage Summary',
+          message:
+            'Issue triage complete for {{variables.owner}}/{{variables.repo}}:\n\n{{node_3.output}}',
+          severity: 'info',
+          position: { x: 1100, y: 250 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4' },
+        { source: 'node_4', target: 'node_5', sourceHandle: 'true' },
+        { source: 'node_4', target: 'node_6', sourceHandle: 'false' },
+      ],
+      variables: {
+        owner: 'my-org',
+        repo: 'my-repo',
+        githubToken: '',
+      },
+    },
+  },
+  {
+    id: 'data-pipeline',
+    name: 'Data Pipeline',
+    description:
+      'Fetch data via HTTP, filter and transform records, aggregate results, summarize with AI',
+    category: 'Data',
+    nodeCount: 7,
+    definition: {
+      name: 'Data Pipeline',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'manual',
+          label: 'Start Pipeline',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          type: 'httpRequest',
+          label: 'Fetch Source Data',
+          method: 'GET',
+          url: '{{variables.dataSourceUrl}}',
+          headers: { Authorization: 'Bearer {{variables.apiToken}}' },
+          position: { x: 300, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'code',
+          label: 'Filter Records',
+          language: 'javascript',
+          code: 'const response = data.body || data;\nconst items = Array.isArray(response) ? response : (response.data || response.results || []);\nreturn items.filter(item => item && typeof item === "object" && item.status !== "archived");',
+          position: { x: 500, y: 150 },
+        },
+        {
+          id: 'node_4',
+          type: 'transformer',
+          label: 'Map Fields',
+          expression:
+            'Array.isArray(data) ? data.map(item => ({ id: item.id, name: item.name || item.title, value: item.amount || item.value || 0, date: item.created_at || item.date })) : []',
+          position: { x: 700, y: 150 },
+        },
+        {
+          id: 'node_5',
+          type: 'code',
+          label: 'Aggregate',
+          language: 'javascript',
+          code: 'const items = Array.isArray(data) ? data : [];\nconst total = items.reduce((sum, i) => sum + (Number(i.value) || 0), 0);\nconst count = items.length;\nconst avg = count > 0 ? total / count : 0;\nreturn { count, total: Math.round(total * 100) / 100, average: Math.round(avg * 100) / 100, items };',
+          position: { x: 900, y: 150 },
+        },
+        {
+          id: 'node_6',
+          type: 'llm',
+          label: 'Summarize Results',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a data analyst. Summarize the aggregated data into a clear, actionable report. Highlight trends, outliers, and key metrics.',
+          userMessage:
+            'Summarize this aggregated data:\n\nTotal records: {{node_5.output.count}}\nTotal value: {{node_5.output.total}}\nAverage value: {{node_5.output.average}}\n\nSample records:\n{{node_5.output.items}}',
+          temperature: 0.4,
+          maxTokens: 2048,
+          position: { x: 1100, y: 150 },
+        },
+        {
+          id: 'node_7',
+          type: 'notification',
+          label: 'Deliver Report',
+          message: '{{node_6.output}}',
+          severity: 'info',
+          position: { x: 1300, y: 150 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4' },
+        { source: 'node_4', target: 'node_5' },
+        { source: 'node_5', target: 'node_6' },
+        { source: 'node_6', target: 'node_7' },
+      ],
+      variables: {
+        dataSourceUrl: 'https://api.example.com/records',
+        apiToken: '',
+      },
+    },
+  },
+  {
+    id: 'scheduled-report',
+    name: 'Scheduled Report',
+    description: 'Gather data on a schedule, analyze with AI, and post results to a webhook',
+    category: 'Scheduling',
+    nodeCount: 5,
+    definition: {
+      name: 'Scheduled Report',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'schedule',
+          label: 'Daily 7AM',
+          cron: '0 7 * * *',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          tool: 'core.search_memories',
+          label: 'Gather Data',
+          args: { query: 'metrics performance status updates', limit: 30 },
+          position: { x: 350, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'llm',
+          label: 'Analyze & Format',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a reporting assistant. Analyze the data and produce a structured daily report in JSON format with sections: "summary", "highlights" (array of strings), "metrics" (key-value pairs), and "actionItems" (array of strings).',
+          userMessage: 'Generate a daily report from this data:\n\n{{node_2.output}}',
+          temperature: 0.4,
+          maxTokens: 3000,
+          responseFormat: 'json',
+          position: { x: 600, y: 150 },
+        },
+        {
+          id: 'node_4',
+          type: 'httpRequest',
+          label: 'Post to Webhook',
+          method: 'POST',
+          url: '{{variables.webhookUrl}}',
+          headers: { 'Content-Type': 'application/json' },
+          body: '{"report": {{node_3.output}}, "generatedAt": "{{variables.timestamp}}"}',
+          bodyType: 'json',
+          position: { x: 850, y: 150 },
+        },
+        {
+          id: 'node_5',
+          type: 'notification',
+          label: 'Report Sent',
+          message: 'Daily report posted to webhook successfully.',
+          severity: 'success',
+          position: { x: 1100, y: 150 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4' },
+        { source: 'node_4', target: 'node_5' },
+      ],
+      variables: {
+        webhookUrl: 'https://hooks.example.com/reports',
+        timestamp: '',
+      },
+    },
+  },
+  {
+    id: 'multi-source-merge',
+    name: 'Multi-Source Merge',
+    description:
+      'Fetch data from 3 APIs in parallel, merge results, transform and synthesize with AI',
+    category: 'AI',
+    nodeCount: 9,
+    definition: {
+      name: 'Multi-Source Merge',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'manual',
+          label: 'Start',
+          position: { x: 100, y: 200 },
+        },
+        {
+          id: 'node_2',
+          type: 'parallel',
+          label: 'Fetch Sources',
+          branchCount: 3,
+          branchLabels: ['Source A', 'Source B', 'Source C'],
+          position: { x: 300, y: 200 },
+        },
+        {
+          id: 'node_3',
+          type: 'httpRequest',
+          label: 'Fetch Source A',
+          method: 'GET',
+          url: '{{variables.sourceAUrl}}',
+          position: { x: 550, y: 50 },
+        },
+        {
+          id: 'node_4',
+          type: 'httpRequest',
+          label: 'Fetch Source B',
+          method: 'GET',
+          url: '{{variables.sourceBUrl}}',
+          position: { x: 550, y: 200 },
+        },
+        {
+          id: 'node_5',
+          type: 'httpRequest',
+          label: 'Fetch Source C',
+          method: 'GET',
+          url: '{{variables.sourceCUrl}}',
+          position: { x: 550, y: 350 },
+        },
+        {
+          id: 'node_6',
+          type: 'merge',
+          label: 'Merge All',
+          mode: 'waitAll',
+          position: { x: 800, y: 200 },
+        },
+        {
+          id: 'node_7',
+          type: 'transformer',
+          label: 'Normalize Data',
+          expression:
+            '({ sourceA: (node_3 && node_3.body) || node_3, sourceB: (node_4 && node_4.body) || node_4, sourceC: (node_5 && node_5.body) || node_5 })',
+          position: { x: 1000, y: 200 },
+        },
+        {
+          id: 'node_8',
+          type: 'llm',
+          label: 'Synthesize Report',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a data synthesis expert. Combine data from multiple sources into a unified analysis. Identify commonalities, discrepancies, and key insights across all sources. Structure your response with clear sections.',
+          userMessage:
+            'Synthesize data from three sources into a unified report:\n\nSource A:\n{{node_3.output}}\n\nSource B:\n{{node_4.output}}\n\nSource C:\n{{node_5.output}}',
+          temperature: 0.5,
+          maxTokens: 4096,
+          position: { x: 1200, y: 200 },
+        },
+        {
+          id: 'node_9',
+          type: 'notification',
+          label: 'Synthesis Complete',
+          message: 'Multi-source synthesis report is ready.',
+          severity: 'success',
+          position: { x: 1450, y: 200 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3', sourceHandle: 'branch-0' },
+        { source: 'node_2', target: 'node_4', sourceHandle: 'branch-1' },
+        { source: 'node_2', target: 'node_5', sourceHandle: 'branch-2' },
+        { source: 'node_3', target: 'node_6' },
+        { source: 'node_4', target: 'node_6' },
+        { source: 'node_5', target: 'node_6' },
+        { source: 'node_6', target: 'node_7' },
+        { source: 'node_7', target: 'node_8' },
+        { source: 'node_8', target: 'node_9' },
+      ],
+      variables: {
+        sourceAUrl: 'https://api.example.com/source-a',
+        sourceBUrl: 'https://api.example.com/source-b',
+        sourceCUrl: 'https://api.example.com/source-c',
+      },
+    },
+  },
+  {
+    id: 'approval-workflow',
+    name: 'Approval Workflow',
+    description:
+      'AI drafts a proposal, submits for human approval, then notifies based on the decision',
+    category: 'Automation',
+    nodeCount: 7,
+    definition: {
+      name: 'Approval Workflow',
+      nodes: [
+        {
+          id: 'node_1',
+          type: 'trigger',
+          triggerType: 'manual',
+          label: 'Submit Request',
+          position: { x: 100, y: 150 },
+        },
+        {
+          id: 'node_2',
+          type: 'llm',
+          label: 'Draft Proposal',
+          provider: 'default',
+          model: 'default',
+          systemPrompt:
+            'You are a professional proposal writer. Draft a clear, concise proposal based on the request. Include: objective, approach, timeline, and expected outcomes. Format it professionally.',
+          userMessage: 'Draft a proposal for this request:\n\n{{node_1.output}}',
+          temperature: 0.6,
+          maxTokens: 3000,
+          position: { x: 350, y: 150 },
+        },
+        {
+          id: 'node_3',
+          type: 'approval',
+          label: 'Review & Approve',
+          approvalMessage:
+            'Please review this AI-drafted proposal:\n\n{{node_2.output}}\n\nOriginal request: {{node_1.output}}',
+          timeoutMinutes: 2880,
+          position: { x: 600, y: 150 },
+        },
+        {
+          id: 'node_4',
+          type: 'condition',
+          label: 'Approved?',
+          expression: 'data === true || data === "approved"',
+          position: { x: 850, y: 150 },
+        },
+        {
+          id: 'node_5',
+          type: 'notification',
+          label: 'Approved',
+          message: 'Your proposal has been approved!\n\nProposal:\n{{node_2.output}}',
+          severity: 'success',
+          position: { x: 1100, y: 50 },
+        },
+        {
+          id: 'node_6',
+          type: 'notification',
+          label: 'Rejected',
+          message:
+            'Your proposal was not approved. Please revise and resubmit.\n\nOriginal request: {{node_1.output}}',
+          severity: 'warning',
+          position: { x: 1100, y: 250 },
+        },
+        {
+          id: 'node_7',
+          tool: 'core.create_memory',
+          label: 'Log Decision',
+          args: {
+            content: 'Proposal decision: {{node_4.output}}. Request: {{node_1.output}}',
+            importance: 6,
+          },
+          position: { x: 1350, y: 150 },
+        },
+      ],
+      edges: [
+        { source: 'node_1', target: 'node_2' },
+        { source: 'node_2', target: 'node_3' },
+        { source: 'node_3', target: 'node_4' },
+        { source: 'node_4', target: 'node_5', sourceHandle: 'true' },
+        { source: 'node_4', target: 'node_6', sourceHandle: 'false' },
+        { source: 'node_5', target: 'node_7' },
+        { source: 'node_6', target: 'node_7' },
+      ],
+    },
+  },
+];
